codebase/requirements.txt
fastapi==0.110.2
google-cloud-aiplatform==1.48.0
google-cloud-storage==2.14.0
google-cloud-vision==3.7.2
google-cloud-documentai==2.27.1
opencv-python==4.9.0.80
scikit-learn==1.5.0
pandas==2.0.3
pillow==10.3.0
PyMuPDF==1.24.2
tensorflow==2.15.1
flask==3.0.3 
openai==1.30.2
uvicorn==0.29.0
imageio==2.9.0
python-multipart==0.0.5
pypandoc==1.13
